# modules/MgkPortlet/MgkPortlet-rest/MgkPortlet-rest-impl/rest-openapi.yaml
info:
    description:
        "A Java client JAR is available for use with the group ID 'com.liferay', artifact ID
        'com.liferay.app.builder.rest.client', and version '1.0.12'."
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "App Builder"
    version: v1.0
openapi: 3.0.1
components:
    schemas:
        Person:
            # @review
            description:
                A person that is a member of the portal.
            properties:
                email:
                    format: email
                    type: string
                firstName:
                    type: string
                id:
                    format: int64
                    readOnly: true
                    type: integer
                lastName:
                    type: string
            type: object
paths:
    "/persons":
        get:
            # @review
            description:
                Retrieves the persons. Results can be paginated.
            operationId: getPersonsPage
            parameters:
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: personId
                  schema:
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Person"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Person"
                                type: array
                    description:
                        ""
            tags: ["Person"]
        post:
            description:
                Creates a new person
            operationId: addPerson
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Person"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Person"
            responses:
                201:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Person"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Person"
                    # @review
                    description:
                        "Person successfully added"
            tags: ["Person"]
        put:
            description: Update person
            operationId: updatePerson
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Person"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Person"
            responses:
                201:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Person"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Person"
                    # @review
                    description:
                        "Person successfully added"
            tags: ["Person"]
        delete:
            description: Delete person
            parameters:
                - in: query
                  name: personId
                  required: true
                  schema:
                      type: integer
            responses:
                "400":
                    description: "Invalid ID supplied"
                "404":
                    description: "Person not found"